feat: Add support for workerd
This allows Cloudflare Workers to import xxhash-wasm as a module.

In the workers runtime you cannot instantiate a WebAssembly module from
an ArrayBuffer. This is to prevent code from code injection attacks.
Instead the runtime can only import a WebAssembly module from a relative
path (it is handled by their bundler).

See https://developers.cloudflare.com/workers/runtime-apis/webassembly/javascript/
for more details on how to import a WebAssembly module in the workers runtime.

To allow workerd to import xxhash-wasm as a module, we add a conditional
export in the package.json file. This allows the workerd runtime to
import the module from the src/index.workerd.js file.

See https://developers.cloudflare.com/workers/wrangler/bundling/#conditional-exports
for more details on how to use conditional exports in a workers project.
