+++ .gitignore
@@ -2,7 +2,9 @@ coverage
 cjs
 esm
 umd
+workerd
 node_modules
 logs
 *.log
 npm-debug.log*
+.yarn/cache
