+++ rollup.config.js
@@ -1,9 +1,10 @@
 import { readFileSync } from "fs";
-import { resolve, dirname } from "path";
-import { fileURLToPath } from "url";
+import { dirname, resolve } from "path";
+import copy from "rollup-plugin-copy";
 import nodeResolve from "rollup-plugin-node-resolve";
-import { swc } from "rollup-plugin-swc3";
 import replace from "rollup-plugin-replace";
+import { swc } from "rollup-plugin-swc3";
+import { fileURLToPath } from "url";
 
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = dirname(__filename);
@@ -35,14 +36,29 @@ const swc_config = JSON.parse(
   readFileSync(resolve(__dirname, ".swcrc"), "utf-8")
 );
 
-export default {
-  input: "src/index.js",
-  output,
-  plugins: [
-    replace(replacements),
-    // The config is necessary, because the plugin overwrites some of the settings,
-    // instead of just falling back to .swcrc
-    swc(swc_config),
-    nodeResolve(),
-  ],
-};
+const plugins = [
+  replace(replacements),
+  // The config is necessary, because the plugin overwrites some of the settings,
+  // instead of just falling back to .swcrc
+  swc(swc_config),
+  nodeResolve(),
+];
+
+export default [
+  {
+    input: "src/index.js",
+    output,
+    plugins,
+  },
+  {
+    input: "src/index.workerd.js",
+    output: { file: "workerd/xxhash-wasm.js", format: "es", sourcemap: true },
+    external: /\.wasm$/,
+    plugins: [
+      ...plugins,
+      copy({
+        targets: [{ src: "src/xxhash.wasm", dest: "workerd" }],
+      }),
+    ],
+  },
+];
